+++ tests/Behat/Page/Admin/GiftCard/UpdatePageInterface.php
@@ -1,0 +1,16 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Tests\Madcoders\SyliusGiftCardPlugin\Behat\Page\Admin\GiftCard;
+
+use Sylius\Behat\Page\Admin\Crud\UpdatePageInterface as BaseUpdatePageInterface;
+
+interface UpdatePageInterface extends BaseUpdatePageInterface
+{
+    public function isCodeEditable(): bool;
+
+    public function getCode(): string;
+
+    public function isInitialAmountEditable(): bool;
+}